Adds Purchase event implementation (#23)
Library updates:
Adds a PurchaseEvent object that is populated by the host app and passed to AttentiveEventTracker
Implements the AttentiveEventTracker class
Updates AttentiveApi to send Event objects generically. Adding more events should need minor if any code changes.
Added a new 'internal' package. This package is for classes that need to be public since they are used across packages in the SDK but should not be used by the host app. An 'internal' package is standard in Android libraries.

Example app updates:
Added a text box in the UI to show the current domain
Sends a Purchase when the purchase is sent