Feature/purchases merge to main (#26)
* Adds purchase UI to the Example app (#21)

* Adds purchase activity to the ui

* more

* Adds EventTracker and Purchase event definition (#22)

AttentiveEventTracker is a singleton that is used to send events

* Adds Purchase event implementation (#23)

Library updates:
Adds a PurchaseEvent object that is populated by the host app and passed to AttentiveEventTracker
Implements the AttentiveEventTracker class
Updates AttentiveApi to send Event objects generically. Adding more events should need minor if any code changes.
Added a new 'internal' package. This package is for classes that need to be public since they are used across packages in the SDK but should not be used by the host app. An 'internal' package is standard in Android libraries.

Example app updates:
Added a text box in the UI to show the current domain
Sends a Purchase when the purchase is sent

* send all ids with purchases (#24)

Without this change, only email/phone/visitorId are sent. This change add the other identifiers.

* Adds OrderConfirmed event (#25)